SKBot 0.0.1

+++ commands/say.js
@@ -2,6 +2,8 @@ const Discord = require("discord.js");
 const ms = require("ms");
 
 module.exports.run = async (bot, message, args) => {
+    if(!message.guild) return
+    console.log("L'utilisateur " + message.author.username + " a dit " + message.content)
     message.delete()
     let MessageToBot = args.join(" ");
     message.channel.send(MessageToBot);
